MSHADE-60 skip dependency of type 'pom' from shading
If one has a dependency which cannot be opened via JarFile we now will at least
throw a verbose Exception with stating the exact file which failed.

I also added an integration test to make sure that the transitive dependencies
pulled in via this pom-dependency gets picked up into the shaded jar.


git-svn-id: http://svn.apache.org/repos/asf/maven/plugins/trunk/maven-shade-plugin@812606 13f79535-47bb-0310-9956-ffa450edef68